+++ .github/workflows/build-and-push.yml
@@ -17,7 +17,7 @@ jobs:
     strategy:
       # max-parallel: 1
       matrix:
-        python-version: ["2.7", "3.8", "3.9", "3.10", "3.11"]
+        python-version: ["2.7", "3.8", "3.9", "3.10", "3.11", "3.12"]
         
     steps:
     - run: echo "Branch $GITHUB_REF"
