+++ jtop/gui/pmem.py
@@ -17,7 +17,6 @@
 
 from os import path
 import curses
-from curses.textpad import rectangle
 # Page class definition
 from .jtopgui import Page
 # Graphics elements
@@ -25,7 +24,6 @@ from .lib.colors import NColors
 from .lib.common import (size_min, unit_to_string, size_to_string, plot_name_info)
 from .lib.linear_gauge import basic_gauge
 from .lib.chart import Chart
-from .lib.button import Button
 from .lib.smallbutton import SmallButton
 
 SWAP_MAX_SIZE = 15
@@ -273,61 +271,4 @@ class MEM(Page):
         # Draw buttons
         self._button_cache.draw(first + height // 2, width - 18, key, mouse, label="clear cache ")
         self.draw_swap_controller(first + height // 2 + 2, width - 18, key, mouse)
-
-    def draw_nv_table(self, start_y, start_x, r_width):
-        columns_title = self.jetson.ram['table'][0]
-        table = self.jetson.ram['table'][1]
-        gpu_val = self.jetson.ram['shared']
-        gpu_val, divider, gpu_unit = size_min(gpu_val, start=self.jetson.ram['unit'])
-        gpu_val_string = str(gpu_val).rstrip('0').rstrip('.')
-        gpu_val_string = "{value}{unit}B".format(value=gpu_val_string, unit=gpu_unit)
-        # Size table
-        r_height = 5 + len(table)
-        # Draw table legend
-        try:
-            rectangle(self.stdscr, start_y, start_x, start_y + r_height, start_x + r_width)
-            self.stdscr.hline(start_y + 2, start_x + 1, curses.ACS_HLINE, r_width - 1)
-            self.stdscr.hline(start_y + 3 + len(table), start_x + 1, curses.ACS_HLINE, r_width - 1)
-        except curses.error:
-            pass
-        # Table
-        size = r_width // len(columns_title) + 1
-        for idx, name in enumerate(columns_title):
-            try:
-                self.stdscr.addstr(start_y + 1, start_x + 2 + idx * size, name, curses.A_BOLD)
-            except curses.error:
-                pass
-            if idx < len(columns_title) - 1:
-                try:
-                    self.stdscr.addch(start_y, start_x + (idx + 1) * size, curses.ACS_TTEE)
-                    self.stdscr.addch(start_y + 1, start_x + (idx + 1) * size, curses.ACS_VLINE)
-                    self.stdscr.addch(start_y + 2, start_x + (idx + 1) * size, curses.ACS_BTEE)
-                except curses.error:
-                    pass
-        for num_row, row in enumerate(table):
-            for num_cell, cell in enumerate(row):
-                if cell == "Size":
-                    value = row[cell]['val']
-                    value, _, unit = size_min(value, start='k')
-                    val_string = str(value).rstrip('0').rstrip('.')
-                    val_string = "{value}{unit}B".format(value=val_string, unit=unit)
-                else:
-                    val_string = row[cell]
-                try:
-                    self.stdscr.addstr(start_y + 3 + num_row, start_x + 2 + size * num_cell, val_string[:size - 2], curses.A_NORMAL)
-                except curses.error:
-                    pass
-                if num_cell < len(columns_title) - 1:
-                    try:
-                        self.stdscr.addch(start_y + 3 + num_row, start_x + (1 + num_cell) * size, curses.ACS_VLINE)
-                    except curses.error:
-                        pass
-        # Total GPU
-        try:
-            self.stdscr.addstr(start_y + 4 + len(table), start_x + 2 + size * (len(columns_title) - 2),
-                               "Shared Tot: {GPU}".format(GPU=gpu_val_string),
-                               (NColors.igreen() | curses.A_BOLD))
-        except curses.error:
-            pass
-        return r_height + 1
 # EOF
