+++ tox.ini
@@ -16,7 +16,7 @@
 
 
 [tox]
-envlist = py{2.7,3.6,3.8,3.9,3.10,3.11}
+envlist = py{2.7,3.6,3.8,3.9,3.10,3.11,3.12}
 skip_missing_interpreters = true
 
 [testenv]
@@ -31,6 +31,7 @@ basepython =
     py3.9: python3.9
     py3.10: python3.10
     py3.11: python3.11
+    py3.12: python3.12
 deps =
     # Reference https://github.com/mgedmin/check-manifest/issues/98
     check-manifest==0.41
